Cleanup service code for plugin_calico
